+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- warn for missing driver
 - remove the injection of bit-js module into the tester environment.
 - add bit-javascript as a dependency and a post install hook.
 - do not show tests output in case of thrown error on commit, use verbose flag to see the error.